+++ models/classes/theme/ConfigurableTheme.php
@@ -20,7 +20,33 @@
 
 namespace oat\tao\model\theme;
 
-interface ConfigurableTheme
+use oat\oatbox\Configurable;
+use oat\tao\helpers\Template;
+
+/**
+ * Class UrlSourceTheme
+ *
+ * Class to easily configure a theme
+ * To use it, declare into tao/theming.conf themes:
+ *
+ * return new oat\tao\model\theme\ThemeService(array(
+ * 'available' => array(
+ *    [...],
+ *    'test' => new \oat\tao\model\theme\UrlSourceTheme(),
+ *    'testConfigured' => new \oat\tao\model\theme\UrlSourceTheme(array(
+ *       'data' => array(
+ *          'logo-url' => 'http://lorempixel.com/400/200,
+ *          'link' => 'http://taotesting.com',
+ *          'message' => 'Tao Platform'
+ *        ),
+ *        'stylesheet' => 'http://tao.dev/tao/views/css/tao-3.css'
+ *     )
+ *  )
+ *  [...]
+ *
+ * @package oat\tao\model\theme
+ */
+class ConfigurableTheme extends Configurable implements Theme
 {
     const THEME_DATA = 'data';
     const THEME_CSS = 'stylesheet';
@@ -30,12 +56,60 @@ interface ConfigurableTheme
     const THEME_DATA_MESSAGE  = 'message';
 
     /**
+     * Get a template associated to given $id
+     *
+     * @param string $id
+     * @param string $context
+     * @return null|string
+     */
+    public function getTemplate($id, $context = Theme::CONTEXT_BACKOFFICE)
+    {
+        switch ($id) {
+            case 'header-logo' :
+                $template = Template::getTemplate('blocks/header-logo.tpl', 'tao');
+                break;
+            case 'footer' :
+                $template = Template::getTemplate('blocks/footer.tpl', 'tao');
+                break;
+            case 'login-message' :
+                $template = Template::getTemplate('blocks/login-message.tpl', 'tao');
+                break;
+            default:
+                \common_Logger::w('Unkown template '.$id);
+                $template = null;
+        }
+        return $template;
+    }
+
+    /**
      * Get options under data key
      * Options to configure header & footer template
      *
      * @return array
      */
-    public function getThemeData();
+    public function getThemeData()
+    {
+        if ($this->hasOption(self::THEME_DATA) && is_array($this->getOption(self::THEME_DATA))) {
+            return $this->getOption(self::THEME_DATA);
+        } else {
+            return [];
+        }
+    }
+
+    /**
+     * Get the url of stylesheet associated to current theme configuration
+     *
+     * @param string $context
+     * @return string
+     */
+    public function getStylesheet($context = Theme::CONTEXT_BACKOFFICE)
+    {
+        if ($this->hasOption(self::THEME_CSS)) {
+            return $this->getOption(self::THEME_CSS);
+        } else {
+            return Template::css('tao-3.css', 'tao');
+        }
+    }
 
     /**
      * Get the logo url of current theme
@@ -43,7 +117,15 @@ interface ConfigurableTheme
      *
      * @return string
      */
-    public function getLogoUrl();
+    public function getLogoUrl()
+    {
+        $data = $this->getThemeData();
+        if (isset($data[self::THEME_DATA_LOGO_URL])) {
+            return $data[self::THEME_DATA_LOGO_URL];
+        } else {
+            return Template::img('tao-logo.png', 'tao');
+        }
+    }
 
     /**
      * Get the url link of current theme
@@ -52,7 +134,15 @@ interface ConfigurableTheme
      *
      * @return string
      */
-    public function getLink();
+    public function getLink()
+    {
+        $data = $this->getThemeData();
+        if (isset($data[self::THEME_DATA_LINK])) {
+            return $data[self::THEME_DATA_LINK];
+        } else {
+            return 'http://taotesting.com';
+        }
+    }
 
     /**
      * Get the message of current theme
@@ -61,5 +151,14 @@ interface ConfigurableTheme
      *
      * @return string
      */
-    public function getMessage();
+    public function getMessage()
+    {
+        $data = $this->getThemeData();
+        if (isset($data[self::THEME_DATA_MESSAGE])) {
+            return $data[self::THEME_DATA_MESSAGE];
+        } else {
+            return '';
+        }
+    }
+
 }
